Add `Model::watch`
